backend/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()  # Charger les variables d’environnement du fichier .env

SECRET_KEY = os.getenv("SECRET_KEY", "supersecretkey")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 30))
ALGORITHM = "HS256"
