backend/requirements.txt
# FastAPI & Serveur ASGI
fastapi==0.110.0
uvicorn==0.29.0

# Base de données avec SQLAlchemy et PostgreSQL
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

# Authentification & JWT
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9  # Pour les formulaires (ex: login)

# Sécurité & Gestion des tokens
pyjwt==2.8.0
cryptography==42.0.7

# Outils pour les tâches & manipulation des dates
pydantic==2.6.4
pydantic-settings==2.1.0  # Gestion des variables d’environnement
python-dotenv==1.0.1  # Pour charger les variables d'environnement (.env)

# Requêtes HTTP & API Clients
httpx==0.27.0
requests==2.31.0

# Dash (Tableau de bord)
dash==2.16.1
dash-bootstrap-components==1.5.0  # Pour améliorer l'UI du Dashboard

# Graphiques et Visualisation de données
plotly==5.20.0
pandas==2.2.1
numpy==1.26.4

# WebSockets (si nécessaire)
websockets==12.0

# Tests et Développement
pytest==8.1.1
pytest-asyncio==0.23.6

# Email
email-validator==2.1.0.post1
